Rename InfoEditor to be more generic
